Clean up User controller.

+++ Zadanie2/TaskBoard/src/main/java/com/TaskBoard/TaskBoard/user/UserService.java
@@ -1,10 +1,14 @@
 package com.TaskBoard.TaskBoard.user;
 
 import com.TaskBoard.TaskBoard.queries.Filter;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.FieldError;
+import org.springframework.web.bind.annotation.RequestBody;
 
 import java.lang.reflect.Field;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Optional;
@@ -13,44 +17,57 @@ import java.util.stream.Collectors;
 @Service
 public class UserService {
     private final UserRepository userRepository;
+    private final UserValidator userValidator;
 
-    @Autowired
-    public UserService(UserRepository userRepository)
+    public UserService(UserRepository userRepository, UserValidator userValidator)
     {
         this.userRepository = userRepository;
+        this.userValidator = userValidator;
     }
-    public User addUser(User user){
-        return userRepository.save(user);
+    public ResponseEntity<String> addUser(User userDetails, BindingResult bindingResult){
+        List<String> getErrors = getResponseEntity(userDetails, bindingResult);
+
+        if(getErrors != null) return ResponseEntity.badRequest().body(String.join("\n",getErrors));
+
+        userRepository.save(userDetails);
+        return ResponseEntity.ok("User has been added");
     }
 
     public boolean existsById(Long id){
         return userRepository.existsById(id);
     }
 
-    public void deleteUserById(Long id) {
+    public ResponseEntity<String> deleteUserById(Long id) {
+        if(!existsById(id))
+            return ResponseEntity.badRequest().body("User with id " + id + " does not exist");
         this.userRepository.deleteById(id);
+        return ResponseEntity.ok("User has been deleted");
     }
 
-    public Optional<User> getUser(Long id) {
+    public ResponseEntity<User> getUser(Long id) {
         Optional<User> user = this.userRepository.findById(id);
-        if(user.isPresent())
-            return user;
-        throw new NoSuchElementException("User is not present");
+        return user.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.badRequest().build());
     }
-    public void updateUser(Long id, User userDetails) {
+    public ResponseEntity<String> updateUser(Long id, User userDetails, BindingResult bindingResult) {
+        if(!existsById(id))
+            return ResponseEntity.badRequest().body("User with id " + id + " does not exist");
+
         User user = userRepository.getReferenceById(id);
+
         user.setName(userDetails.getName());
         user.setSurname(userDetails.getSurname());
         user.setEmail(userDetails.getEmail());
         userRepository.save(user);
+
+        return ResponseEntity.ok().body("user has been updated");
     }
 
-    public List<User> getUsers() {
-       return this.userRepository.findAll();
+    public ResponseEntity<List<User>> getUsers() {
+       return ResponseEntity.ok(this.userRepository.findAll());
     }
 
-    public List<User> getFilteredUsers(Filter userFilter) {
-        return this.userRepository.findAll().stream()
+    public ResponseEntity<List<User>> getFilteredUsers(Filter userFilter) {
+        List<User> userList =  this.userRepository.findAll().stream()
                 .filter(user -> {
                     try {
                         Field field = User.class.getDeclaredField(userFilter.getFieldName());
@@ -66,6 +83,21 @@ public class UserService {
                     return false;
                 })
                 .collect(Collectors.toList());
+        if(userList.isEmpty())
+            return ResponseEntity.badRequest().build();
 
+        return ResponseEntity.ok(userList);
+    }
+
+    private List<String> getResponseEntity(User userDetails, BindingResult bindingResult) {
+        userValidator.validate(userDetails,bindingResult);
+        if (bindingResult.hasErrors()) {
+            List<String> errorMessages = new ArrayList<>();
+            for (FieldError fieldError : bindingResult.getFieldErrors()) {
+                errorMessages.add(fieldError.getDefaultMessage());
+            }
+            return errorMessages;
+        }
+        return null;
     }
 }
